Create V5-core.py
Tried using a separate core over a thread however the start up is problematic. MPV starts before window is created and means that it doesn't show properly work around was to press a button once the ui was created and that solved the problem. 

Button response was still not great, think there is an occasional issue with running mpv on a separate core. 

Overall threading works better so going that route. this is just here as a reference.